requirements.txt
Django==2.2.11
django-bootstrap-datepicker-plus==3.0.5
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pep257==0.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.1
